refactor(solidlsp): Phase 5 - extract SymbolQueryService and location request classes
Move SymbolLocationRequest and subclasses, primitive request_* methods
(request_definition/implementation/references), higher-level methods
(request_referencing/containing/defining/implementing_symbols,
request_symbol_at_location, request_container_of_symbol), and helpers
(_request_symbol_at_location, _refine_implementing_symbol,
_get_preferred_definition, etc.) into src/solidlsp/ls_symbol_query.py.

Introduce SymbolQueryService as a deep collaborator that receives narrow
callbacks to the owner facade for policy hooks and delegates to
OpenDocuments (for open_file/_open_file_context) and DocumentSymbolRepository
(via request_document_symbols).

SolidLanguageServer facade is now thin for these concerns: retains public
request_* as delegators, keeps all override hooks (_get_preferred_definition,
_send_*_request, _create_text_document_position_params, _pre_open_*,
_wait_for_cross_file_*, etc.), and hosts nested *LocationRequest classes
plus module-level aliases for Svelte adapter compatibility and pickle/backcompat.

All public signatures and adapter hook names unchanged. Lint and type-check
clean. Full test suite shows only pre-existing failures (C++/Unreal + minor).
Characterization tests (Phase 0-4 baselines) remain fully green.
